rework the parser to generate a list of "custom types" over all storyboards, with a merged list of segues per type
